+++ src/backend-single.c
@@ -78,8 +78,9 @@ void laik_single_exec(Laik_Data* d, Laik_Transition* t, Laik_TransitionPlan* p,
             assert(to > from);
 
             laik_log(1, "Single reduce: "
-                        "from %ld, to %ld, elemsize %d, base from/to %p/%p\n",
-                     from, to, d->elemsize, (void*) fromBase, (void*) toBase);
+                        "from %lld, to %lld, elemsize %d, base from/to %p/%p\n",
+                     (long long int) from, (long long int) to,
+                     d->elemsize, (void*) fromBase, (void*) toBase);
 
             memcpy(toBase, fromBase, (to-from) * fromMap->data->elemsize);
         }
